scripts/monitor_sql/requirements.txt
sense-hat
requests
mysql-connector
RTIMULib